Improve rendering of graphics points
Graphics points are no longer being rendered for the master node set. In place of the master node set we are instead rendering only the nodes that are not part of a node group. This means that we no longer have to double render nodes that are part of both the master node set and a node set group.