Add combo box for selection mode
This commit adds a combo box to the widget UI allowing the user to choose the selection mode assigned to the left mouse button. Currently the choices are _Exclusive_ or _Additive_.

Since the SceneSelection handler defines the selection mode on mouse press event, I've had to extend the SceneSelection class to define a custom mouse press event.