Add coordinate field combo boxes
This commit adds two combo boxes to the plugin GUI. These allow the user to specify which of the coordinate fields in the input file should be used for the point-cloud and mesh objects. Previously we were just using the field names created by the Auto Segmentation step, which works well enough when the two steps are connected but not when this plugin uses an EX file from an alternative source.