Add hashing of input data
This commit introduces hashing of the point cloud input data to the step settings.

Before loading the point cloud into the points region we now generate a hash of the point cloud input data and check it against the hash of the input data that was used last time this step was opened. The previous output file will be reloaded if the current input data matches the previous input data.

The hash of the point cloud input data is saved to the step settings when closing the step.